+++ source/app/blueprints/graphql/permissions.py
@@ -26,7 +26,10 @@ from flask import request
 from app.util import get_case_access_from_api
 from app.iris_engine.access_control.utils import ac_get_effective_permissions_of_user
 from app.iris_engine.access_control.utils import ac_fast_check_current_user_has_case_access
-from app.business.errors import PermissionDeniedError
+
+
+class PermissionDeniedError(Exception):
+    pass
 
 
 def _deny_permission():
